Added security and privacy to delete and edit contents

+++ Project/classes/Ijdb/Controllers/Joke.php
@@ -8,16 +8,19 @@
 
 namespace Ijdb\Controllers;
 
+use Amghezi\Authentication;
 use \Amghezi\DatabaseTable;
 use \Amghezi\Controller;
 
 class Joke extends Controller {
 	private $jokesTable;
 	private $authorsTable;
+	private $authentication;
 	
-	public function __construct(DatabaseTable $jokesTable , DatabaseTable $authorsTable) {
+	public function __construct(DatabaseTable $jokesTable , DatabaseTable $authorsTable, Authentication $authentication) {
 		$this -> jokesTable = $jokesTable;
 		$this -> authorsTable = $authorsTable;
+		$this -> authentication = $authentication;
 	}
 	
 	public function list() {
@@ -32,12 +35,16 @@ class Joke extends Controller {
 				'jokeText' => $joke['jokeText'],
 				'jokeDate' => $joke['jokeDate'],
 				'name' => $author['name'],
-				'email' => $author['email']
+				'email' => $author['email'],
+				'authorId' => $author['id'],
 			];
 		}
 		
+		$user = $this->authentication->getUser();
+		
 		$variables['jokes'] = $jokes;
 		$variables['totalJokes'] = $this -> jokesTable -> total();
+		$variables['userId'] = $user['id'] ?? null;
 		
 		return $this->return('Jokes', 'jokes', $variables);
 	}
@@ -52,17 +59,27 @@ class Joke extends Controller {
 			$_SESSION['post'] = null;
 			unset($_SESSION['post']);
 		}
-		$this->jokesTable->delete($_POST['id']);
+		
+		$user = $this->authentication->getUser();
+		$joke = $this->jokesTable->findById($_POST['id']);
+		
+		if($user['id'] == $joke['authorId']) {
+			$this->jokesTable->delete($_POST['id']);
+		}
+		
 		header('LOCATION: /joke/list');
 	}
 	
-	public function addOrEdit() {
+	public function getForm() {
 		// Add action
 		$title = 'Add Joke';
+		
+		$user = $this->authentication->getUser();
 		
 		// Edit action
 		if(isset($_GET['id'])) {
 			$variables['joke'] = $this->jokesTable->findById($_GET['id']);
+			$variables['userId'] = $user['id'];
 			$title = 'Edit joke';
 		}
 		
@@ -71,14 +88,16 @@ class Joke extends Controller {
 	
 	public function save() {
 		$joke = $_POST['joke'];
-		$author = $this->authorsTable->find('email', $_SESSION['username'])[0];
+		$author = $this->authentication->getUser();
 		
 		// If the action is and edit,the author is set ,
 		// Else, we should assign it to the logged-in user
 		$joke['authorId'] = $this->jokesTable->findById($joke['id'])['authorId'] ?? $author['id'] ?? 7;
 		
-		// Do edit or insert action
-		$this->jokesTable->save($joke);
+		// Do edit or insert action if the joke is assigned to the user
+		if($joke['authorId'] == $author['id']) {
+			$this->jokesTable->save($joke);
+		}
 		
 		// In either case of the action, redirect the user
 		header("LOCATION: /joke/list");
